+++ packages/story-editor/src/components/canvas/navLayer.js
@@ -17,51 +17,17 @@
 /**
  * External dependencies
  */
-import { memo, useCallback } from '@web-stories-wp/react';
-import { __ } from '@web-stories-wp/i18n';
-import { ContextMenu } from '@web-stories-wp/design-system';
+import { memo } from '@web-stories-wp/react';
 import Proptypes from 'prop-types';
-import styled from 'styled-components';
 
 /**
  * Internal dependencies
  */
-import { useQuickActions } from '../../app/highlights';
-import DirectionAware from '../directionAware';
 import Carousel from '../carousel';
-import { useLayout } from '../../app';
 import { CarouselArea, HeadArea, Layer, PageMenuArea, Z_INDEX } from './layout';
-import PageMenu from './pagemenu';
-
-const MenusWrapper = styled.div`
-  position: absolute;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: space-between;
-  z-index: 9999;
-  pointer-events: auto;
-  min-height: 100%;
-`;
+import PageSideMenu from './pageSideMenu';
 
 function NavLayer({ header }) {
-  const { hasHorizontalOverflow } = useLayout(
-    ({ state: { hasHorizontalOverflow } }) => ({ hasHorizontalOverflow })
-  );
-  const quickActions = useQuickActions();
-
-  /**
-   * Stop the event from bubbling if the user clicks in between buttons.
-   *
-   * This prevents the selected element in the canvas from losing focus.
-   */
-  const handleMenuBackgroundClick = useCallback((ev) => {
-    ev.stopPropagation();
-  }, []);
-
-  const showQuickActions =
-    !hasHorizontalOverflow && Boolean(quickActions.length);
-
   return (
     <Layer
       pointerEvents="none"
@@ -70,25 +36,7 @@ function NavLayer({ header }) {
     >
       <HeadArea pointerEvents="initial">{header}</HeadArea>
       <PageMenuArea>
-        <DirectionAware>
-          <MenusWrapper>
-            {showQuickActions && (
-              <ContextMenu
-                isInline
-                isAlwaysVisible
-                isIconMenu
-                disableControlledTabNavigation
-                groupLabel={__(
-                  'Group of available options for selected element',
-                  'web-stories'
-                )}
-                items={quickActions}
-                onMouseDown={handleMenuBackgroundClick}
-              />
-            )}
-            <PageMenu />
-          </MenusWrapper>
-        </DirectionAware>
+        <PageSideMenu />
       </PageMenuArea>
       <CarouselArea pointerEvents="initial">
         <Carousel />
